+++ web/client/components/security/SecurityPopupDialog.jsx
@@ -6,17 +6,31 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-import React from 'react';
+import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import {Glyphicon} from 'react-bootstrap';
+import {Glyphicon, InputGroup, ControlLabel, FormGroup, FormControl as FormControlRB} from 'react-bootstrap';
 
 import Modal from '../misc/Modal';
 import Message from '../I18N/Message';
-import Button from '../layout/Button';
+import _Button from '../layout/Button';
 import FlexBox from '../layout/FlexBox';
 import Text from '../layout/Text';
 import Spinner from '../layout/Spinner';
+import uuidv1 from 'uuid/v1';
+import tooltip from '../misc/enhancers/tooltip';
+import withDebounceOnCallback from '../misc/enhancers/withDebounceOnCallback';
+import localizedProps from '../misc/enhancers/localizedProps';
+import { getCredentials } from '../../utils/SecurityUtils';
 
+const FormControl = localizedProps('placeholder')(FormControlRB);
+
+function _InputControl({ onChange, value, ...props }) {
+    return <FormControl {...props} value={value} onChange={event => onChange(event.target.value)}/>;
+}
+
+const InputControl = withDebounceOnCallback('onChange', 'value')(_InputControl);
+
+const Button = tooltip(_Button);
 
 function SecurityPopupDialog({
     show,
@@ -24,20 +38,39 @@ function SecurityPopupDialog({
     disabledClose,
     onCancel,
     onConfirm,
+    onClear,
     titleId,
-    disabled,
     confirmId,
     variant,
     loading,
-    children,
     preventHide,
-    titleParams
+    titleParams,
+    service,
+    debounceTime,
+    maxLength,
+    showPassword,
+    setShowPassword
 }) {
+
+    const [creds, setCreds] = useState(getCredentials(service?.protectedId));
+    const disabled = !(creds.username && creds.password);
+
+    function handleCancel() {
+        onCancel();
+    }
 
     function handleHide() {
         if (!loading && !preventHide) {
-            onCancel();
+            handleCancel();
         }
+    }
+    function handleClear() {
+        onClear();
+        setCreds({});
+    }
+    function handleConfirm() {
+        onConfirm(creds);
+        setCreds({});
     }
 
     if (!show) {
@@ -54,15 +87,67 @@ function SecurityPopupDialog({
                         {titleId ? <Message msgId={titleId} msgParams={titleParams} /> : null}
                     </Text>
                     {showClose && onCancel &&
-                        <Button square borderTransparent disabled={disabledClose} onClick={onCancel}>
+                        <Button square borderTransparent disabled={disabledClose} onClick={handleCancel}>
                             <Glyphicon glyph="1-close"/>
                         </Button>
                     }
                 </FlexBox>
-                {children}
+                <FormGroup>
+                    <InputGroup>
+                        {service?.url}
+                    </InputGroup>
+                </FormGroup>
+                <FlexBox inline>
+                    <FormGroup style={{
+                        flex: 1,
+                        padding: "0px 5px 0px 0px"
+                    }}>
+                        <ControlLabel>
+                            <Message msgId="securityPopup.username" />
+                        </ControlLabel>
+                        <InputControl
+                            name={`username_${uuidv1()}`}
+                            value={creds.username}
+                            debounceTime={debounceTime}
+                            onChange={(username) => setCreds({...creds, username})}
+                            maxLength={maxLength}
+                        />
+                    </FormGroup>
+                    <FormGroup style={{
+                        flex: 1,
+                        padding: "0px 5px 0px 0px"
+                    }}>
+                        <ControlLabel>
+                            <Message msgId="securityPopup.pwd" />
+                        </ControlLabel>
+                        <InputGroup style={{width: "100%"}}>
+                            <InputControl
+                                name={`password_${uuidv1()}`}
+                                autoComplete="new-password"
+                                type={showPassword ? "text" : "password"}
+                                value={creds.password}
+                                debounceTime={debounceTime}
+                                onChange={(password) => setCreds({...creds, password })}
+                                maxLength={maxLength}
+                            />
+                            <InputGroup.Addon>
+                                <Button
+                                    tooltipId={showPassword ? "securityPopup.hide" : "securityPopup.show" }
+                                    onClick={() => {setShowPassword(!showPassword);}}>
+                                    <Glyphicon glyph={!showPassword ? "eye-open" : "eye-close"}/>
+                                </Button>
+                            </InputGroup.Addon>
+                        </InputGroup>
+                    </FormGroup>
+                    <FormGroup style={{alignContent: "flex-end"}}>
+                        <Button onClick={handleClear} tooltipId="securityPopup.remove" >
+                            <Glyphicon glyph="trash"/>
+                        </Button>
+                    </FormGroup>
+                </FlexBox>
                 <FlexBox centerChildrenVertically gap="sm">
                     <FlexBox.Fill />
-                    <Button disabled={disabled || loading} variant={variant} onClick={() => onConfirm()}>
+                    <Button disabled={disabled || loading} variant={variant} onClick={handleConfirm}>
                         <Message msgId={confirmId} />
                         {loading ? <>{' '}<Spinner /></> : null}
                     </Button>
@@ -79,9 +164,9 @@ SecurityPopupDialog.defaultProps = {
     showClose: false,
     onCancel: () => {},
     onConfirm: () => {},
+    onClear: () => {},
     titleId: '',
     errorId: '',
-    disabled: false,
     disabledClose: false,
     loading: false,
     preventHide: true,
@@ -93,8 +178,8 @@ SecurityPopupDialog.propTypes = {
     show: PropTypes.bool,
     onCancel: PropTypes.func,
     onConfirm: PropTypes.func,
+    onClear: PropTypes.func,
     titleId: PropTypes.string,
-    disabled: PropTypes.bool,
     loading: PropTypes.bool,
     preventHide: PropTypes.bool,
     confirmId: PropTypes.string,
